server/requirements.txt
fastf1
panda
numpy
scikit-learn
matplotlib